requirements.txt
appdirs==1.4.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.11
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==22.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
oauthlib==3.1.1
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==2.4.0
pymongo==3.11.4
python-decouple==3.4
python-docx==0.8.10
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
tzdata==2022.2
urllib3==1.26.6
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
